refactor(orderbook): change doubles to bigdecimal to closely follow api spec

+++ src/main/kotlin/com/example/orderbook/api/dto/OrderDTO.kt
@@ -1,8 +1,10 @@
 package com.example.orderbook.api.dto
 
+import java.math.BigDecimal
+
 data class OrderDTO(
     val side: String,
-    val quantity: Double,
-    val price: Double,
+    val quantity: BigDecimal,
+    val price: BigDecimal,
     val currencyPair: String
 )
