+++ src/main/kotlin/com/example/orderbook/model/Order.kt
@@ -1,5 +1,7 @@
 package com.example.orderbook.model
 
+import java.math.BigDecimal
+
 enum class OrderSide {
     BUY, SELL
 }
@@ -7,7 +9,7 @@ enum class OrderSide {
 data class Order(
 //    val id: String, TODO(): come back to this e.g. cancelling orders as a stretch goal
     val side: OrderSide,
-    var quantity: Double, // TODO(): Big decimal for sats
-    val price: Double,
+    var quantity: BigDecimal,
+    val price: BigDecimal,
     val currencyPair: String,
 )
